requirements.txt
aiogram==2.9.2
aiohttp==3.6.2
async-timeout==3.0.1
asyncpg==0.20.1
attrs==19.3.0
Babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
gino==1.0.1
idna==2.10
multidict==4.7.6
python-dotenv==0.13.0
pytz==2020.1
SQLAlchemy==1.3.18
yarl==1.4.2
